requirements.txt
PySide2==5.15
Flask==2.1.2
flask-restx==0.5.1
Werkzeug==2.1.2
Flask-Classful
pyngrok
python-dotenv